osc-core: shrink ConfigIdentity::firmware_version u16 → u8 to match dxl 2.0 ping wire

+++ firmware/lib/core/src/services/dxl/dispatcher.rs
@@ -361,7 +361,7 @@ impl Dispatcher<'_> {
                 error: StatusError::OK,
                 status: PingStatus {
                     model: identity.model_number,
-                    fw_version: identity.firmware_version as u8,
+                    fw_version: identity.firmware_version,
                 },
             },
         );
